requirements.txt
flask
openai
line-bot-sdk
